Create new release 3.8.4

+++ Backup/Dto/Interfaces/RemoteUploadDtoInterface.php
@@ -11,9 +11,9 @@ interface RemoteUploadDtoInterface
 
     public function setIsAutomatedBackup(bool $isAutomatedBackup);
 
-    public function getTotalBackupSize(): int;
+    public function getTotalBackupSize(): float;
 
-    public function setTotalBackupSize(int $totalBackupSize);
+    public function setTotalBackupSize(float $totalBackupSize);
 
     public function getFilesToUpload(): array;
 
@@ -23,7 +23,7 @@ interface RemoteUploadDtoInterface
 
     public function setUploadedFiles(array $uploadedFiles = []);
 
-    public function setUploadedFile(string $uploadedFile, int $fileSize);
+    public function setUploadedFile(string $uploadedFile, float $fileSize);
 
     public function getIsOnlyUpload(): bool;
